redo refactoring changes
